feat(DTFS2-7052): renaming enums/geospatial.ts to enums/geospatialCountries.ts
