+++ src/constants/enums.ts
@@ -1,9 +1,9 @@
 import * as FALLBACK_TO_LEGACY_DATA from './enums/fallbackToLegacyData';
-import * as GEOSPATIAL from './enums/geospatial';
+import * as GEOSPATIAL_COUNTRIES from './enums/geospatialCountries';
 import * as PRODUCTS from './enums/products';
 
 export const ENUMS = {
   PRODUCTS: PRODUCTS.QueryParamProductsEnum,
   FALLBACK_TO_LEGACY_DATA: FALLBACK_TO_LEGACY_DATA.FallbackToLegacyDataEnum,
-  GEOSPATIAL_COUNTRIES: GEOSPATIAL.GeospatialCountriesEnum,
+  GEOSPATIAL_COUNTRIES: GEOSPATIAL_COUNTRIES.GeospatialCountriesEnum,
 };
